Properly get coverage from all tests.
* Add COVERAGE=1 Makefile option to build libraft with coverage.
* Refactor CFFI module to use the existing shared object in an
  out-of-line configuration, so there's no need to build and re-build
  it.
* Makefile adjustments to build the module CFFI as needed.
* Remove raft_get_snapshot_entry_idx, a prototype for an undefined
  function that breaks CFFI.
* Add a gcov Makefile target to fetch both unit tests and integration
  test coverage in one shot.
